pyproject.toml
[project]
name = "jutra"
version = "0.1.0"
description = "jutra: conversational future-self backend (ADK + Gemini 3 + Firestore + MCP for LiveKit)"
requires-python = ">=3.11,<3.13"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [{ name = "jutra hackathon team" }]

dependencies = [
  "google-adk>=1.31.0",
  "google-genai>=1.0.0",
  "google-cloud-firestore>=2.19.0",
  "google-cloud-secret-manager>=2.21.0",
  "google-cloud-aiplatform>=1.71.0",
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.32.0",
  "pydantic>=2.9.0",
  "pydantic-settings>=2.6.0",
  "python-multipart>=0.0.12",
  "mcp>=1.2.0",
  "httpx>=0.27.0",
  "orjson>=3.10.0",
  "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "pytest-asyncio>=0.24.0",
  "pytest-cov>=5.0.0",
  "hypothesis>=6.115.0",
  "ruff>=0.7.0",
  "mypy>=1.13.0",
  "types-requests",
]

[project.scripts]
jutra = "jutra.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["jutra"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
  "E", "F", "W",      # pycodestyle + pyflakes
  "I",                 # isort
  "B",                 # bugbear
  "UP",                # pyupgrade
  "SIM",               # simplify
  "C4",                # comprehensions
  "RET",               # return
  "PTH",               # pathlib
]
ignore = [
  "E501",  # line-length handled by formatter
  "B008",  # FastAPI Depends in default args
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["B", "SIM"]
# OCEAN Big-Five uses single-letter trait names (O, C, E, A, N) by convention.
"jutra/personas/**" = ["E741"]

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers --tb=short"
markers = [
  "slow: slow tests (network / live Gemini calls)",
  "live: requires live GCP credentials",
]

[tool.mypy]
python_version = "3.11"
strict = true
exclude = ["tests/", "demo_data/"]
